refactor: splict `Backend` into submodules
