+++ src/storage/backend/mod.rs
@@ -1,0 +1,9 @@
+//! The backend storage.
+
+mod backend_impl;
+
+pub use backend_impl::{Backend, BackendBuilder};
+
+#[cfg(test)]
+#[allow(clippy::unwrap_used, clippy::indexing_slicing)]
+mod test;
